Sectioned debug extensions off to the debug class
