+++ Nyxa3D/Nyxa3D.vcxproj.filters
@@ -113,4 +113,9 @@
       <Filter>Source Files\Callback</Filter>
     </None>
   </ItemGroup>
+  <ItemGroup>
+    <Text Include="Todo.inl">
+      <Filter>Source Files</Filter>
+    </Text>
+  </ItemGroup>
 </Project>
